fix type-check
