+++ edge-functions/multi-brand/package.json
@@ -18,6 +18,7 @@
     "swr": "^1.2.1"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.1",
     "@types/node": "17.0.14",
     "@types/react": "17.0.38",
     "autoprefixer": "^10.4.2",
